// frontend/src/lib/backend/apiV1/schema.d.ts
export interface components {
    schemas: {
        Empty: Record<string, never>;
        RsDataEmpty: {
            resultCode: string;
            msg: string;
            data: components["schemas"]["Empty"];
        };
        PostCommentModifyReqBody: {
            content: string;
        };
        PostCommentDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            /** Format: int64 */
            postId: number;
            /** Format: int64 */
            authorId: number;
            authorName: string;
            content: string;
        };
        RsDataPostCommentDto: {
            resultCode: string;
            msg: string;
            data: components["schemas"]["PostCommentDto"];
        };
        PostModifyReqBody: {
            title: string;
            content: string;
            published?: boolean;
            listed?: boolean;
        };
        PostWithContentDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            /** Format: int64 */
            authorId: number;
            authorName: string;
            authorProfileImgUrl: string;
            title: string;
            published: boolean;
            listed: boolean;
            content: string;
            actorCanModify?: boolean;
            actorCanDelete?: boolean;
        };
        RsDataPostWithContentDto: {
            resultCode: string;
            msg: string;
            data: components["schemas"]["PostWithContentDto"];
        };
        MemberModifyMeReqBody: {
            nickname: string;
        };
        MemberDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            nickname: string;
            profileImgUrl: string;
        };
        RsDataMemberDto: {
            resultCode: string;
            msg: string;
            data: components["schemas"]["MemberDto"];
        };
        PostWriteReqBody: {
            title: string;
            content: string;
            published?: boolean;
            listed?: boolean;
        };
        PostCommentWriteReqBody: {
            content: string;
        };
        MemberLoginReqBody: {
            username: string;
            password: string;
        };
        MemberLoginResBody: {
            item: components["schemas"]["MemberDto"];
            apiKey: string;
            accessToken: string;
        };
        RsDataMemberLoginResBody: {
            resultCode: string;
            msg: string;
            data: components["schemas"]["MemberLoginResBody"];
        };
        MemberJoinReqBody: {
            username: string;
            password: string;
            nickname: string;
        };
        PageDtoPostDto: {
            /** Format: int32 */
            currentPageNumber: number;
            /** Format: int32 */
            pageSize: number;
            /** Format: int64 */
            totalPages: number;
            /** Format: int64 */
            totalItems: number;
            items: components["schemas"]["PostDto"][];
        };
        PostDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            /** Format: int64 */
            authorId: number;
            authorName: string;
            authorProfileImgUrl: string;
            title: string;
            published: boolean;
            listed: boolean;
        };
        PostStatisticsResBody: {
            /** Format: int64 */
            totalPostCount: number;
            /** Format: int64 */
            totalPublishedPostCount: number;
            /** Format: int64 */
            totalListedPostCount: number;
        };
        MemberWithUsernameDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            nickname: string;
            profileImgUrl: string;
            username: string;
        };
        PageDtoMemberWithUsernameDto: {
            /** Format: int32 */
            currentPageNumber: number;
            /** Format: int32 */
            pageSize: number;
            /** Format: int64 */
            totalPages: number;
            /** Format: int64 */
            totalItems: number;
            items: components["schemas"]["MemberWithUsernameDto"][];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}